Add UseStorage trait to CashTrickler
Add UseStorage trait to CashTrickler. It defaults to false. When true the cash trickle will be added to the resource storage. It obeys the same rules that harvesters dumping resources do. The cash tick will show the amount granted. If silos are needed it will either show nothing or a partial tick depending on how much free space is available.